+++ src/studio/fs.h
@@ -41,7 +41,7 @@ typedef enum
 typedef bool(*ListCallback)(const char* name, const char* info, s32 id, void* data, bool dir);
 typedef void(*AddCallback)(const char*, AddResult, void*);
 typedef void(*GetCallback)(GetResult, void*);
-typedef void(*OpenCallback)(const char* name, const void* buffer, size_t size, void* data);
+typedef void(*OpenCallback)(const char* name, const void* buffer, s32 size, void* data);
 
 typedef struct FileSystem FileSystem;
 
@@ -52,8 +52,8 @@ void fsAddFile(FileSystem* fs, AddCallback callback, void* data);
 void fsGetFile(FileSystem* fs, GetCallback callback, const char* name, void* data);
 bool fsDeleteFile(FileSystem* fs, const char* name);
 bool fsDeleteDir(FileSystem* fs, const char* name);
-bool fsSaveFile(FileSystem* fs, const char* name, const void* data, size_t size, bool overwrite);
-bool fsSaveRootFile(FileSystem* fs, const char* name, const void* data, size_t size, bool overwrite);
+bool fsSaveFile(FileSystem* fs, const char* name, const void* data, s32 size, bool overwrite);
+bool fsSaveRootFile(FileSystem* fs, const char* name, const void* data, s32 size, bool overwrite);
 void* fsLoadFile(FileSystem* fs, const char* name, s32* size);
 void* fsLoadFileByHash(FileSystem* fs, const char* hash, s32* size);
 void* fsLoadRootFile(FileSystem* fs, const char* name, s32* size);
@@ -69,7 +69,7 @@ bool fsExists(const char* name);
 void* fsReadFile(const char* path, s32* size);
 bool fsWriteFile(const char* path, const void* data, s32 size);
 bool fsCopyFile(const char* src, const char* dst);
-void fsGetFileData(GetCallback callback, const char* name, void* buffer, size_t size, u32 mode, void* data);
+void fsGetFileData(GetCallback callback, const char* name, void* buffer, s32 size, u32 mode, void* data);
 void fsOpenFileData(OpenCallback callback, void* data);
 void fsOpenWorkingFolder(FileSystem* fs);
 bool fsIsDir(FileSystem* fs, const char* dir);
